move xlsx serializer to utils
